Minor clean to example codes
